Added authorization handlers & tests
- Added three new handlers:
1. /auth/user_id ---> returns the status of the requested person object.
If the requested object is logged in, the handler will only reply with
the information if the requested user belongs to the currently logged in
user (so you can't get the status on other users)
2. /auth/login --> login handler that will see if you have already
logged in. If so, it redirects to the main page, otherwise it will
redirect you to a Google login page.
3. /auth/logout --> similar to the login handler, the logout handler
will check if you are logged in. If so, it will log you out and redirect
you to the login page and otherwise it will redirect you immediatly.

With handler 2 and 3, there are still some known issues. I don't know
how redirects will affect mobile devices (since it's an url redirect)
and we decided to create our own login page whilst the login/logout
handlers currently use google's page.