Updated handlers & unittests after merge
-Updated all authorization handlers (status, login, logout) to have them
use the new global function that gets the person object that belongs to
the user that is currently logged in.
- Updated all unittests to the changes made in the handlers. All
negative tests had to be removed since there are no more errors being
raised anywhere; all unittests are positive but still achieve a 100%
code coverage.

- Changed status handler response type: if the user is not logged
in/unknown in the application, only the is_logged_in field is returned
(with False). Otherwise, the person_id and is_application_admin is also
returned.
- Removed id paramters from all auth handlers since they will now always
ask for the data of the currently logged in user.