get_current_person update
-Update get_current_person_update function to take in a model reference
parameter that will be used to fetch the person object that belongs to a
logged in user, rather than a static reference to model.Employee.
-Changed return types to a dictionary containing both the person object
(if found) and also a boolean that indicates whether there is a user
logged in. This way, if there can be distinction between the two
scenario's in which no person object is returned, rather than always
having False/None returned.