Changed setattr workings and unittests
- Removed global function to get a property reference since this wasn't
need anymore
- Added function in BaseTestClass to intaintiate a server with a dummy
handler that can be used to test model workings without the need to send
HTTP requests (so you can still put model objects)
- Changed unit tests to actually test the permission system for setting
the declaration properties
- Changed both the __setattr__ function and the
handle_custom_property_set_operation function within declaration. The
setattr function now checks if the key is custom, and only checks if it
is allowed to be set with the current class name if it is custom (so
non-custom properties will be set normally). The
handle_custom_property_set_operation function now retrieves the old
value for the custom attibute to be setted, and sets it to this default
value if there is no permission to set it. Otherwise, it sets it to the
new value.
- Changed Test data creator to start off with the class_name property.
This should now always be set first.
- Changed type of permission values from property references to strings
(since the key param is always a string).