Added supervisor action support
Added the possibility for supervisors to either approve, decline or
forward a declaration. Details:
- Fixed missing parameter in controller to allow navigation
- Set the button code inside the controller, eliminating a previous todo
- Created/updated two generic methods for the three actions to be
called: showMessageInputForDeclarationAction and handleSentDeclaration:
---handleSentDeclaration---
- This is updated and now takes in a request_data parameter instead of a
comment, allowing a new supervisor to also be passed.
- Updated the error in handleSentDeclaration to use the server's error
message instead of a hard coded one.
- When the request is done, the dialog (see bellow) will now
automatically close.
---showMessageInputForDeclarationAction---
- New function that (much like the showMessage in standardFunctions.js)
takes in several parameters to show a dialog.
- Takes in a title, a button value (title), the message body, a callback
to be executed when the button within the dialog is clicked and an
action type (approve, forward or decline). When decline is passed,
filling out a comment will be mandatory and the request cannot be sent
untill one has been given.
- Added a text input field within the dialog to allow supervisors to
comment. As described in the previous point, this is mandatory when
declining a declaration. When approving or forwarding and no comment is
filled out, it wil NOT be sent as part of the request body (filtered out
client-side).
- This function is called in the clicked callbacks of the three action
buttons on the sent declataion detail page. The forward button will
always create a message where the currently selected supervisor is
incorperated. If the selected supervisor is the same as the one
currently asisgned on the declaration, an error message is shown instead
(note that no selection takes place within the dialog; the selection
happens at the sent declaration detail page).