Major update to sent decl. actions
*IMPORTANT*
Re-formatted the code in the sentDeclarationDetailsController completely
so it will be hard to check the code differences if you follow git's
markup system.

- Revamped the entire way sent declarations are handled for
approving/declining/forwarding:
- Changed message input shown when a button is clicked to include an
error area that can display a faulty input (for instance, declining has
to have a comment)
- Added support for future tasks: HR approve/decline. Added a date
picker that is hidden by default but can be shown if the
should_show_date_selector is true (default is false).
- Added reading of date (mm-YYYY) function on the submit button click.
This only happens if the should_show_date_selector parameter is set to
true. If it is false, the date will not be appended in the data
dictionary that is passed to a callback.
- Fixed issue with error dialog not showing when another dialog was
already open.
- Fixed issue with un-usable JSON data being received when an error
occured after a sent declaration action. Though the main issue is still
there (posted data is being appended after the error response text), a
workaround is currently in place and holding up. This will break though,
whenever an object with a '}' character is returned.
- Moved the form input checks to the action button clicks in the detail
view, rather than the function that makes the server request or shows
the input dialog.
- Removed need for action identifier in
showMessageInputForDeclarationAction function (because of the previous
point).
- Added the new supervisor's id to the request data when forwarding a
declaration.
- Changed some minor commentary errors, text labels, text formatting and
added some CSS entities to accomodate the error messages.
- Added some lines to the ForwardDeclarationHandler in
declaration_handler.py to indicate what sort of data is being sent to
the back-end.