Added HR action functionalities for sent decl.
- Changed HTML of the sentDeclarationDetails.html:
- All elements that should be hidden for HR, are hidden and same
goes for supervisor elements.
- Changed the supervisor selection box (for forwarding when user is a
supervisor) to be a rugalar td element for HR.
- Also changed the action controls for both roles: the buttons are now
in seperate divisions that will be hidden/shown based on the logged
in person's role.
- Minor changes to back-end:
- Fixed two defects in the decline/approve handlers for HR. There was
one call in both these handlers to the is_declaration_state function
that passed a set object, rather than a string for the class name to
check (see difference).
- Added the missing statement for the will_be_payed_out_on property
for declarations when a declaration is being approved by HR.
- Renamed existing button handlers for supervisor actions to distinguish
them from the HR handlers.
- Fixed the $state.go defect: they now reference the right (updated)
template.
- Added two handlers for approving and declining declarations for HR.
Approving must always pass a date. Comments are in both cases optional.
- Changed the date picker tool:
- Years are being build dynamically from the current year: the first
option will be the current year and the last option will always be 10
years from the current year. This means that this has never to be
updated in code.
- The month selector now always starts at the current month.