Code review changes
- Two changes to the senDeclarationDetails HTML:
- Changed how the supervisor is displayed from two td elements being
hidden to one td elements with two label elements
- Changed how the action (approve/forward/decline) buttons work from
different button to the same buttons with the same callbacks. Moved
userclass checks to javascript controller.
- Changed the sentDeclarationDetailsController: removed button handlers
so that now there are only 3 handlers. These handlers will check what
user rights the person have and call the appropriate function based on
that. These functions are what used to be in the seperate handlers (no
changes to these functions)
- Also updated one thing according to Mikel's wishes: Removed the
approve button completely when the total price of the declaration is
bigger than the max price the supervisor can approve, and a label will
be shown when this is the case.