Updated button hiding for invalid conditions
- Added another check based on Mikel's suggestions: if a supervisor has
no supervisor himself, the forward button is disabled and a label is
shown in the same position as the other label (price label).
- Had to change html & css structure since a margin is needed to
seperate the components if a label is shown. This is now done
dynamically with the angular ng-style directive and a function that
calculates how big the margin should be.